"Add language switch feature to website theme"
This commit introduces a language switch feature in the website theme component. A new function "getLanguage" has been added which gets the language from local storage, defaulted to 'en'. The switching function 'toggleActiveFlag' sets the language in local storage. Also, the 'activeFlag' state is initialized with the stored or default language.  The changes also include text localization for "Developer" in English.