"Added language toggle and updated translations"
This commit introduces a language toggle functionality to the themes component. A new property, 'setUpdatedLanguage', is added to the ThemesProps interface to allow the setting of a new language. This is used in a new useEffect hook, which sets the updated language to the active flag when it changes.

In the App component, 'setUpdatedLanguage' and 'updatedLanguage' states are added, with the latter being used to change the language through i18n based on user's choice.

Translations in 'en.json' and 'ptBR.json' have also been modified to reflect more expressive welcome message.