Add internationalization to About page section titles
In this commit, the static strings used in the About page section's subtitles were replaced with a function call to the translation function `t()`. This change allows different languages to be presented depending on the selected language. New translations were added to the en.json and ptBr.json files for "My Skills" and "Experience & Education".