Remove unused code from React-based application
This commit removes the file 'dist/assets/index-0297365d.js' which was previously imported but wasn't being used anymore. The file was removed to reduce the overall size of the application and make it load faster. The removal doesn't affect the functionality of the application as it was confirmed that no other files, modules, or components used its contents. The deletion also helps maintain a clean project structure by removing unnecessary and unused code.